Encrypt passwords using bcrypt, keep sha256 for backwards compatibility
The number of rounds can be specified by BCRYPT_ROUNDS environment variable. The default value is 14, as it takes 600 ms on my machine.
As we currently only support bcrypt and plain sha256 hashes, we can distinguish them by the leading '$' of bcrypt hashes.
Upgrading hashes on login is not yet implemented.